Extract IpPoolCell so we can use it in other places
